apps/api/pyproject.toml
[tool.poetry]
name = "amethyst-api"
version = "0.1.0"
description = "Amethyst API Server"
authors = ["Arko Bhattacharya <arko@fask.ai>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
amethyst-engine = {path = "../../packages/engine", develop = true}
sse-starlette = "^3.0.3"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

